CRA-024 test refactoring with separate util class [HM,SC]
